fix: sync code

Created-by: tiantianok
Commit-by: tiantianok
Merged-by: TimZhou
Description: fix: sync code

See merge request: Cangjie/cangjie_deveco_plugins!4

+++ tools-extension/cjlint/src/main/java/com/huawei/deveco/cjlint/CjlintCodeCheckCmd.java
@@ -266,10 +266,14 @@ public class CjlintCodeCheckCmd extends CodeCheckCmd {
                     throw new ExecutionException(message);
                 }
                 editCheckResult();
-                FileUtils.delete(new File(cjOutPath));
+                if (StringUtils.isNotEmpty(cjOutPath)) {
+                    FileUtils.delete(new File(cjOutPath));
+                }
             }
             // 清空本项目临时存放的自定义CJLint配置文件
-            FileUtils.deleteDirectory(new File(cacheCJLintConfigPathStr));
+            if (StringUtils.isNotEmpty(cacheCJLintConfigPathStr)) {
+                FileUtils.deleteDirectory(new File(cacheCJLintConfigPathStr));
+            }
             // 根据code-linter.json5中的files和ignore配置，筛选检查结果resultList
             filterResultListByCodeLinterJson5();
             if (returnEmptyResult) {
